openstack-ardana: use workspace symlink for artifacts
The `junit` and `archiveArtifacts` pipeline plugins don't support
absolute paths, which means they can only take in files that are
in the Jenkins `WORKSPACE` folder. Using a dedicated workspace for
every job run, which is cleaned up at the end, and creating symlinks
there from the shared workspace solves this problem.