Always cleanup the workspace
The workspace cleanup was moved to the prep script, but it was still
only called when github_pr was set, possibly making other runs using an
automation repo with patches that were being tested in an earlier run.

Fix the prep script to not raise an error without $github_pr.

Make it explicit that the directory being emptied is $WORKSPACE. Replace
rsync from empty with rm.