openstack-ardana: fix ardana-qe-tests
This change update ardana-qe-tests to use the same structure used on
other playbooks since [1]. Besides that, it changes the way test results
are processed, instead requiring each test to grep the log for results,
now the test results are taken from the subunit output for all tests.

It also introduces a new variable `ardana_qe_test_timeout` which can be
used to define a timeout for running the test. By default it is set to
120 minutes.

1. https://github.com/SUSE-Cloud/automation/pull/2766