openstack-ardana: switch Gerrit CI to demo input model for Cloud 9
Monasca is going to be broken for a while in Cloud 9, until all
the needed monasca packages are ported to Rocky. In the meantime,
we still need to maintain a working CI, so disabling monasca
for all Cloud 9 Ardana CI jobs (switching to `demo` as an input model,
which has monasca disabled) is a fair compromise.

This commit also disables all Ardana Cloud9 CI jobs that are using
input models including monasca, for the same reasons.