openstack-ardana: fix URLs for failing downstream jobs
Using try-catch blocks to display the URLs of downstream jobs
in the Blue Ocean only works if the downstream job is passing.
When failing, the try-catch block will not set the needed variables
and the resulting "undefined variable" failure just eclipses the
real cause of the failure.

The correct approach is to use `propagate: false` when launching
downstream builds, which will complete successfully regardless of
the downstream build result, then check the result in the upstream
job.