Fix devstack job in ci.opensuse.org
Previous commit allowed the devstack jobs to use some jenkins
plugins to store and report the test results.
However, there were several mistakes, mainly for rushing things :(.

The junitxml dependency has to be added to the VM were devstack is
running. Also, the if to check for the file can be avoided since we want
to fail the build if it doesn't exist.

Signed-off-by: aojeagarcia <aojeagarcia@suse.com>