Add want_ipv6
Which will add a bunch of ipv6 networks:

  fd00:0:0::1::0/64 - public
  fd00:0:0::2::0/64 - fixed
  fd00:0:0::3::0/64 - admin
  fd00:0:0::4::0/64 - storage
  fd00:0:0::5::0/64 - ceph
  fd00:0:0::6::0/64 - ironic
  fd00:0:0::7::0/64 - sdn

NOTE: we could just pick a prefix (ie above is mostly zeros), or
randomise each time. But not sure if that'll play havic with tests.

This uses a 64 prefix so there is enough room for SLAAC, which we currently
plan to use for the admin node, at least. To keep IPv6 and IPv4
automation as similar as possible, the octet of ipv4 and last ipv6
block in most cases will match.

Ie: .100 == :100 or .254 == :254.

Because IPv6 addresses need to be wrapped when used directly as URLs
there are a bunch variable and a new function to make that happen.

A seperate network.json (scripts/lib/ipv6/netmaks.json) was added to adjust
to make patching and managing IPv6 easier. However, this file is apart
of the automation repo, not coming from crowbar, so will need to be maintained.

Currently this has only been tested up until getting the admin node up,
so the IPv6-ification is still a while away.