mkcloud: Switch to a pcid compatible guest cpu
Now that we have a specific CPU type for AMD based hosts,we
no longer need to default to a host cpu that is 11 years old in order
to find a lowest common denominator. With Haswell we are exposing pcid,
which is highly recommended to mitigate to cost of Meltdown mitigiations
on the guest. Haswell is matching the oldest Intel CPU in our CI
cluster and is meanwhile 5 years old. noTSX disables transactional
memory extension, which we don't need.