Make obs_scm continue to work with local git path
In 44b3beeaf8663d46467fb3945ddc3d73c65ac452
from https://github.com/openSUSE/obs-service-tar_scm/pull/251
obs_scm was changed to reject anything that is not a remote URL like a
local path. We use a local path to build a test package for git commits
that are not yet merged. Work around this change by setting the
environment variable that is used to bypass this for testing obs_scm.