openstack-ardana: fix create os resources tasks on ardana_qe_tests role
For creating openstack resources (projects, users, etc) we use ansible
modules which requires the shade library. For that, shade is installed on
the test virtual environment and the ansible python interpreter is set to
use python from the virtual environment.

Using `~` on the virtual enviroment path causes ansible to fail with
`OSError: [Errno 2] No such file or directory`. So this path replaces
`~` with the full path.