input-model-generator: use clm_model in service template
Remodel `clm_model` as a service template parameter,
and use it to select automatically the service groups
for which the CLM service components are configured
depending on the `standalone` or `integrated` value.

This improvement removes the need to have a dedicated
service template for the each of the two scenarios:
standalone and integrated deployer.