openstack-ardana: get cloud media build number from motd
The cloud media build number is always added to the motd by the
`setup_zypper_repos` role. Other roles relies on different methods to
get the cloud media build number, this is very error prone as depending
of the cloudsource there are different methods to get the cloud media
build number and there are cases where cloudsource is not available
(running tempest job).

With this change, instead of applying the same logic from `setup_zypper_repos`
to get the media build number, other roles read that info from the motd
as it was previously added.