input-model-generator: disable additional LVM volumes
The virtual Ardana cloud deployments currently disable
all provisioning of LVM volumes by means of the `skip_disk_config`
mechanism. One consequence of this is that services use the
root partition for all their files while the additional
`volume-group` entries in the generated disk model are ignored.

This commit removes this additional `volume-group` entries and
relies only on the root partition the size of which is controlled
only by the ECP openstack flavor associated with the virtual server
instances. The immediate effect is a huge (~50%) reduction in the
storage resources required to instantiate virtual cloud environments
associated with generated input models.