input-model-generator: fix external network support
Using the `neutron.l3_agent.external_network_bridge` tag to
indicate the provider network used as a backend by the neutron
external network has been broken since cloud8 [1]. Using this
tag will effectively supply no provider network attributes to
the external neutron network, which in turn causes neutron
to automatically select one of the tenant networks to act as
a backend for the external network.

This commit fix the generated external neutron network
configuration to provide explicit provider attribute values
and to model the external network as a flat provider network.

The commit also changes the generated names used for the external neutron
networks to match the name used by the `ardana-input-model` input models
(e.g. `ext-net`, `ext-net1` and so on), because the implementation of many
of the QA test automation scripts is still based on this assumption.

[1] https://bugzilla.suse.com/show_bug.cgi?id=1100583