Fix domain shut off test
This patch fixes a test for when a virsh domain is shut down.
The old text was "shut.off", but virsh has since changed to "shut off".
This also fixes doing a snapshot, which first shuts the domain down.
Without this fix, snapshotting will fail after it asks the domain to shut
down.