input-model-generator: separate management and neutron networks
An issue occurs intermittently when the management network is
co-attached with one of the neutron networks on the same interface
[1]. To avoid that, the input model generator scenarios are updated
such that the MANAGEMENT network is no longer attached to the same
interface as the neutron external network (EXTERNAL-VM) or the neutron
VLAN tenant network (TENANT-VLAN).

To preserve the current association with bare-metal hardware interfaces,
the bond interfaces in the `compact-bond` and `full-spread` interface
templates also had to be split, to accommodate the increase in required
number of interfaces. New templates were created for this, while the
"bond" ones are preserved.

[1] https://bugzilla.suse.com/show_bug.cgi?id=1067482