openstack-ardana: remove SDK SP3 repo hack
As a temporary solution to the problem of python-devel/python-base
package version mismatch between the SLES12-SP4 base image and the
released SLES12-SP4-SDK GM, the SP3 SDK-Updates repo was also included
before installing python-devel.

The root cause of the problem was identified as an incorrectly
built SLES12-SP4 image and was corrected by including the
`snapshot-SP4` repositories from SLE-12 update projects instead
of the `standard` ones [1], hence this hack is no longer required.

[1] https://build.suse.de/project/meta/home:fmccarthy:ardana-images