Automatically perform recheck when outdated
Automatically perform a recheck when a dependency or other change causes
the results for this change to be outdated. This also means that any
running jobs will be cancelled and rescheduled.

This hardcodes the branch to job which isn't ideal, however, our future
product testing will likely not use this.

Fixes SCRD-5848