openstack-ardana: merge tempest and QA tests jobs
The current implementation does not allow runnig QA tests if any tempest
test (or the tempest stage) failed. To overcome that, the tempest and QA
test jobs have been merged into one job, which makes possible to select
multiple tempest and QA tests and they will be executed independently if the
previous has failed or not.