ssl: Try not to use insecure certs
In the real world we don't recommend using self-signed certs. We need to
test the real-world use case in which certificates are not generated by
crowbar but instead installed by the operator.

This patch turns off the generate_certs and insecure flags in some of
the proposals and instead generates the key pair and installs it in the
system CA bundle. Things break if we turn if off for the database and
rabbitmq proposals, so holding off on those for now. Not bothering with
crowbar and ceph.