openstack-ardana: generate pipeline results dynamically
Generates pipeline job results based on the date returned
by the Jenkins pipeline REST API using a python script.
These results are saved as a Jenkins job artifact and,
in the case of Gerrit jobs, also reported back to Gerrit.

The `jenkins-job-pipeline-report` script takes in a job
name and build number and generates a hierarchy of pipeline
stages indicating their status and URLs.

Implements SCRD-5820