Base CCP job periodic
This adds a minimum CCP job in a pipeline, running a shell
script from the socok8s repository, with configurable branche.

Adding the pipeline brings the ability to ALWAYS run the cleanup
steps.

This script creates and configures 1 SES, 3 CaaSP, and a deploy
node into an existing openstack environment
(by choice we'll be using engcloud-cloud-ci, to not overlap with
existing ardana resources).

This job will run periodically (once per day).