openstack-ardana-gating: add deploy-and-update test (SCRD-3705)
This patch updates the Ardana gating jobs to also run the
`cloud-ardana[8|9]-job-std-3cp-devel-staging-updates-x86_64`
job in addition to the `cloud-ardana[8|9]-job-std-min-x86_64`
job previously executed.

In order to do this in parallel, the ardana-gating job also
had to be converted to a pipeline job, because it is not
currently possible to trigger a Jenkins pipeline jobs from a
Jenkins multi-job.

As the URLTrigger Jenkins plugin is not yet compatible with
Jenkins pipeline jobs, an additional `cloud-ardana[8|9]-gating-trigger`
job is used to trigger the gating job on URL changes.