ssl: Don't use insecure clients
python-requests generates a lot of logs about insecure endpoints when we
use the --insecure option for openstackclients. Now that in
develcloud7/8/9 the certificates are added to the trust bundle, we don't
actually need to use --insecure, so ensure we are using the secure
version when we can so that we have cleaner logs.