openstack-ardana-heat: volume mountpoint mapping
Mapping volumes to server mountpoints doesn't work in OpenStack [1].
The way we enforced that was by controlling the order in
which volumes get attached to servers through heat `depends_on`
hints.

However, this doesn't seem to work anymore. With Pike, we very
frequently get the root partition mounted as /dev/vdb and a regular
volume mounted as /dev/vda.

This patch attempts a different approach, serializing the creation
of server, volumes and volume attachment resources by linking them
with `depends_on` restrictions. The time it takes to create a volume
will increase the chance that volumes are attached in the right order.

[1] https://bugs.launchpad.net/nova/+bug/1004328