switch to unified github context string: suse/cloud
The previous context string "suse/mkcloud" was specific
to mkcloud. So it would be misleading to let jobs testing
other deployment tools report a status using this context
string.

On the other hand adding new context strings and requiring
them as mandatory for a merge would mean that on most PRs
some checks will never be run, because not all PRs run with
all deployment tools

So this new context string
- must have a success status which is mandatory for a merge
- will be used for the status of the default test (PR build)
  of this PR.
  Hint: this is the "standard" job, that is defined in the
  github-pr yaml in:
  https://github.com/SUSE-Cloud/automation/blob/master/scripts/github_pr/github_pr_cloud.yaml#L5
  The "standard" is not added as suffix to the context, while
  all other build names (keys on the same level in the yaml) are
  appended as suffix to the base context.
- is the base string for all additional (optional and mandatory)
  contexts that are required (this is already the case).
  e.g.
    old: suse/mkcloud/stable
    new: suse/cloud/stable