Disable cinder backup if swift is not enabled
All tempest tests in tempest.api.volume.test_volumes_backup and
tempest.api.volume.admin.test_volumes_backup are failing because
cinder-backup is trying to use swift as a backend.
However, because of we are using Python3, swift can not be enabled since
is still not supported.

This patch disable cinder backup in case that swift is not enabled per
 https://review.openstack.org/#/c/75073/

Signed-off-by: aojeagarcia <aojeagarcia@suse.com>