openstack-ardana: lower timeout for stack creation
Sometimes (eg. due to some issue on ECP) the stack gets stuck on CREATING
state until it times out which by default is 1 hour. However creating a heat stack
takes in average 3 minutes and if it takes much longer it will probably
not be crated successfully.

This change sets the heat stack creation timeout to 8 minutes (should be
more than enough to create the stack) so we can throw the error and try
again faster.