SCRD-6075 Add versioning support for designate services
For Cloud9 we want to switch to using the designate worker/producer
services than the deprecated zone/pool managers. To support this we
define a new versioned feature, designate_worker_producer, which is
true for Cloud9 deployments.

This new versioned feature setting is checked when defining the CORE
service components list, and we generate either worker or pool and
producer or zone entries accordingly in the list.

NOTE: This change should not be landed until the corresponding patch
to the ardana/designate-ansible.git repo that fixes worker/producer
deployment for designate in Cloud9, with Gerrit ChangeID
I647ff48154d4af3e1c400b6ff8bc973a56ab81bd.