openstack-ardana: use dedicated slots for gating jobs
Change the lockable resources allocation strategy to use
dedicated slots for Ardana IBS gating jobs. This approach
guarantees that the cloud environments that are left running
after a gating job failure are not not destroyed by other CI
jobs of less importance and are therefore available for debugging.

This should significantly shorten the amount of time needed to
identify and fix the issues affecting the gating jobs.