openstack-ardana: use generated input models for cloud8
Switch the Ardana Cloud8 periodic and gating jobs to using generated
input models, to be similar to their Cloud9 counterparts.

This also includes switching the cloud9 'demo' input model job and creating
a similar job for cloud8. This minimalistic job is important, because
it's the fastest cloud deployment currently available to us (NOTE:
the 'demo' input model is an input model with one controller-deployer
node, one SLES compute node and with MML services disabled. Aside
from MML services, octavia and freezer also need to be disabled
because they are tightly coupled with Monasca.)