ipv6: Add network details to routed.cloud
If want_ipv6 is set to > 0, set the routed cloud on the mkch* hosts to
use the prefix from 99c199e928c7 for the admin network. Based on the
admin network prefix, set the public network prefix so that they
overlap. Create a random IPv6 prefix for the ironic network.