Run only smoke tests on devstack jobs
It turns out that running the same regex for tempest in our devstack
jobs we never were able to have a green job.

Having jobs without tests passing is useless, so until we will able to
have it green it's better to switch back to the previous state where we
only ran the smoke tempest tests.

In addition, we improve the logging of the job leveraging the
subunit-trace command to show the logs in the console

Signed-off-by: aojeagarcia <aojeagarcia@suse.com>