don't configure git user name and email to non-human
Generally git's user.email and user.name should only ever be set to
reflect actual humans.  Setting to something else opens the risk of
making commits which are assigned to a non-human.  In the rare case
where a non-human setting is required, that should be done by the
invoking environment.