add devel:languages:python:backports and Cloud:OpenStack:Master repos
devstack expects to be able to install a bunch of dependencies, e.g.
haproxy and Python packages like python-psutil, etc.

Since devstack is an upstream tool, it needs to be able to function
correctly outside of SUSE, so we have to use a combination of
repositories from the SLE product and from OBS.