don't force installation onto untested distributions
When installing onto an untested distro release, it's safer default
behaviour to allow devstack to bail with a warning to the user.

In the case where someone is working on adding support for a new
distro release, they can take advantage of the newly added ability to
clone a modified fork of devstack, by ensuring that their fork adds
the distro under test to the list of approved distros.