input-model-generator: configurable generated IP values
Modified the logic generating IP addresses and subnet prefixes
for input models to also cover the Crowbar case, which expects
192.168.124.0/24 to be the admin network prefix, the .10 IP
address to be the admin node and the rest of IP addresses to
start with .80.

To support this feature, a set of global variables were added
to the input model generator ansible role, which can be overridden
by specifying different values in the scenario files.