openstack-ardana-gating: make update optional
The Ardana IBS gating job runs two validation tests in parallel:
one that deploys an Ardana cloud based on the staging cloud
media (DC8S, DC9S), the other that deploys an Ardana cloud based on the
non-staging media (DC8) and then updates the cloud installation to
the staging media.

In some circumstances, both staging and non-staging builds are
broken (e.g. by a submitrequest breaking a package in the non-staging
project) and are subsequently fixed by patches that land in the
staging area. In these cases, the deploy-and-update job will not
pass and needs to be manually circumvented to unblock the CI/CD
pipeline.

This change allows the gating job to be (re)started with the update
validation step disabled.