Lower CPU model for Intel
Running `mkcloud` in a VM on the ECP is currently only possible if the
hypervisor is at least Haswell based. A large fraction of our hardware
is older than that and `libvirt` fails with

   libvirt.libvirtError: the CPU is incompatible with host CPU: Host
      CPU does not provide required features: fma, movbe, fsgsbase,
      bmi1, avx2, smep, bmi2, erms, invpcid

on those hypervisors. This change lowers the default CPU model to
SandyBridge.

Also adjust the test fixtures.

Signed-off-by: Nicolas Bock <nicolas.bock@suse.com>