openstack-ardana: use automation_update
With the shared workspace mechanism out of the way,
the Ardana pipeline jobs can be modified to use the
Crowbar CI approach of managing a shared
SUSE-Cloud/automation repository git clone on every
Jenkins agent, instead of creating a new git clone
every time a job runs.

This is accomplished by calling the same script
that is used by the Crowbar CI - `update_automation`,
which basically creates or updates the shared git
clone - but with two important differences, which are
described below.

The Ardana CI jobs make use of a `git_automation_repo`
and `git_automation_branch` pair of parameters, by default
pointing to the SUSE-Cloud/automation repository and master
branch, which can be modified to point to a custom fork,
branch and/or commit number instead.

The need for these parameters comes from the way that Jenkins
pipeline jobs work: the Jenkins job configuration needs a git
repository location from which to load the job pipeline definition
(i.e. the Jenkinsfile). For most cases, this is the
SUSE-Cloud/automation master branch, but we need to support
these additional use-cases:

1. an Ardana CI contributor wants to test some changes available
in a custom git fork, branch or commit, without publishing them as
a GitHub pull request. For this to be possible, the Jenkins jobs
need to be triggered manually and pointed to the custom git fork,
branch or commit.

2. an Ardana CI contributor wants to run custom tests on an
automation repository GitHub pull request. This should be as
easy ad triggering Jenkins jobs and pointing them to a PR number.

3. ultimately, the "CI for CI" - the CI that validates automation
repository GitHub pull request - needs to run automated testing
against an open pull request

The second thing that needs to be done differently from the Crowbar
CI is the way the shared git clone is actually used. The Ardana CI
implementation is based on the idea that the back-end scripts (e.g.
ansible, python) can be run independently from Jenkins and directly
from a local copy of the automation git repository. To preserve this
assumption and allow these scripts to run and modify the contents of
the automation repository freely, as if executed from a local copy,
most of the Ardana Jenkins jobs cannot simply share the automation
repository git clone, they need to create a copy of it in their
workspace, which can be modified independently of the master git clone.

Implements: https://jira.suse.de/browse/SCRD-4983
Partially implements: https://jira.suse.de/browse/SCRD-4307