openstack-ardana-heat: remove shared workspace
The `openstack-ardana-heat` Jenkins job is the only job still
actively using the concept of a "shared workspace" - sharing
a file location on the Jenkins agent between two job runs.
The requirement comes from the fact that the generated heat
template file describing the Ardana cloud environment needs to
be passed from the `openstack-ardana` down to the `openstack-ardana-heat`
job that will use it to instantiate the cloud environment.

Using a text (multi-line string) job parameter achieves the same
result, while at the same time removing the need to share a file
location between upstream and downstream jobs. This is the first
step towards removing the concept of a shared workspace and also
making the `openstack-ardana-heat` job a more generic
job that handles heat stack creation/deletion, which could be
reused by other projects as well (e.g. to create the SES cluster).

Another advantage is the fact that a user may now also trigger the
job manually to create a heat stack, by simply dumping the heat
template contents into the text input.