remove ssh driver from ironic proposal
SSH drivers were deprecated in Ironic long time ago and proposal validation fails
with these. In addition, agent_* drivers could be enabled if Swift is deployed.

SOC9+ uses new "hardware types" instead of classic drivers and these should have
decent defaults so no need to set anything for that. This part could be extended
in the future if custom values are needed.