openstack-ardana: JENKINS-52638 bug workaround
There's a bug in the Lockable Resources plugin that sometimes
causes a Null value to be returned by the `lock` function called
to reserve a resource [1][2]. This has been known to happen for
example when the Jenkins server is restarted, or when a resource
that was previously manually reserved is released and immediately
reserved by one of the jobs waiting in queue.

We don't yet have a fix for this bug, nor a way around it, but we
can at least abort the jobs immediately if this bug is detected,
rather than having them continue to run without a locked resource.

[1] https://issues.jenkins-ci.org/browse/JENKINS-52638
[2] https://issues.jenkins-ci.org/browse/JENKINS-43002