openstack-ardana: use closure to reserve resources
Reserving Lockable Resources for Ardana pipeline jobs involves more
than just calling the build-in 'lock' step, which leads to duplicating
code in a lot of places - basically everywhere a resources needs to be
reserved.

This commit changes that by implementing a closure that builds on the
default 'lock' step behavior and making it available via the Groovy
shared library.