IPv6: Allow mkvlan support a netmask for ipv6
Currently the mkcloud mkvlan function hardcodes a netmask
of 24. This patch uses the $defaultnetmask if $want_ipv6
is set, which defaults to 64.
Otherwise it remains the 24 default.

This patch also adds an $admingw like we do for ironic and
admin ip. Which makes it easier to manage IPv4 and IPv6
addresses on the mkcloud host.