openstack-crowbar: ansible wrapper for qa_crowbarsetup.sh
Implements an ansible wrapper that can be used to run
`qa_crowbarsetup.sh` commands on the Crowbar admin node.
Its implementation is largely based on `mkcloud`, but using
ansible specific mechanisms:
- the `mkcloud.config` file is generated based on a template
that uses ansible variables (which have default values that
can be overriden by the Jenkins job parameters. Currently,
only a minimum number of `mkcloud.config` variables have
generated values, while the rest are only included as comments.
- the crowbar_deploy role synchronizes the `scripts` folder
to the admin node, generates the `mkcloud.config` file, then
calls `qa_crowbarsetup.sh` on the admin node.

The wrapper is currently used only to install Crowbar on the
admin node.

In addition to this, the following pipeline stages have been
implemented as ansible playbooks:
* bootstrap-admin-node: resizes the root partition on the admin node and
  performs a number of minor tasks required to prepare the admin node for
  Crowbar installation
* bootstrap-crowbar-nodes: resizes the root partition on the cloud nodes