openstack-ardana-image-update: share images (SCRD-6094)
We're currently using two sets of SLES CI images in the
ECP cloud, due to the fact that we cannot create public
images with the `cloud-ci` ECP user, which is used to
automate image update, and the image membership support
was thought to be broken in Newton (i.e. because
`openstack --os-cloud engcloud-cloud-ci image set --accept <image>`
didn't have the desire effect).

After further investigations, it turns out that the explanation
for the above failure was simple: we just had to use a
project UUID instead of the project name when sharing
the image.

This commit makes the image update automation use shared
images by default and also removes the need to have OpenStack
cloud configuration files preconfigured for the other
projects with which the image needs to be shared.

The only pre-requisite is that the OpenStack user used by the
automated job (`cloud-ci` in our particular case) also needs
to be associated with the other projects.