qa_crowbarsetup: Add "want_l3_ha"
This is useful to deploy a cloud with the upstream solution for l3 HA.
This requires https://github.com/crowbar/crowbar-openstack/pull/2005
which adds l3 HA support to Crowbar.

Note: If this PR and the crowbar PR are merged, another PR is needed
to enable l3 HA by default for the gating jobs.