openstack-ardana: Add script to deploy ardana without jenkins
This change adds a bash script that mimics the steps followed by the
openstack-ardana job to deploy ardana on ECP.

The intent of the script is to make easy deploying ardana on ECP from
the workstation (without the need of jenkins). This change also
adds a library that provides functions similar to the openstack-ardana
jenkins job steps, which can be called to execute the each step
individually.