openstack-ardana-gerrit: fix pipeline job report
The `jenkins-job-pipeline-report` script that is generating
the execution summary report of a pipeline job was not detecting
downstream jobs properly. This patch fixes the script such that
any number of downstream jobs can be triggered from a pipeline
stage and they will all be included in the report regardless of
their position in the set of steps comprising the stage.