jjb: add timeout for testbuild-pr-trigger
Without the timeout a build that got stuck
(this happened) would block the following
builds. So rather exit the one that is stuck
to allow for a new run.

The timeout should be less than the interval
used for the job. Currently the interval is
10 minutes and the timeout of 500 seconds
is about 8 minutes.
Normal runtime of this job is 2 - 3 minutes.