openstack-ardana: set nova cpu_model for virtual deployments
Running live migration tests on ardana deployments on ECP occasionally
fails due to CPU incompatibility (e.g. one compute is scheduled to a sandy
bridge host and another to a westmere host).

This change sets the cpu model to Westmere on the deployed ardana computes,
making sure that it always possible to do livre migration on the
deployed ardana cloud.