Move manila-share to controller (SCRD-7771)
Move manila-share to controller for new CI template. This will apply to only in
c8. According to the ticket, we make this decision in order to be
consistent with SOC. As for c9, we deprecate manila-share in the input model.
Due to manila-api and manila-share residing on the same host, both ansible
roles can use reference from each other and doesn't have to use MNL-SHR host
explicitly.

1. Move manila-share to controller
2. Add filter to enable manila-share only on c8