ipv6: Unset http_proxy in libvirt setuphost
setuphost is run by CI for every job, and it uses zypper to make sure
certain packages are installed, except that routed.cloud script for
mkch* hosts sets http_proxy unilaterally to an address that won't exist
until the libvirt bridge is created. Unset the variable for the zypper
call.