openstack-crowbar: deploy batch scenario stage (SCRD-7369)
Implements the crowbar proposal stage of the `openstack-crowbar`
pipeline. Until a crowbar batch scenario file is generated to
match the generated input model, one of the existing scenario
files needs to be supplied as an input parameter.

A change is required to `crowbar_register`, to set the node aliases
according to the crowbar batch scenario files used by the legacy
Crowbar CI (`compute-kvmX` instead of `computeX`).