ipv6: Be absolutely clear about steps to execute
It turns out the steps to execute are literally passed to mkcloud, which
results in behaviour that isn't conducive to testing anything at all.
Specify all of the steps for the job we want to be run.

Drop instcrowbar for the time being since Crowbar installation fails on
IPv6 admin nodes.