openstack-ardana: std-lmm input model scenario and jobs
Creates a new input model generator scenario, `std-lmm`,
which is similar to the `standard` scenario, with the only
difference that the LMM services are configured in a separate
cluster.

This scenario moves LMM services away from the controller nodes,
in an attempt to make the CI jobs more reliable by elliminating
the instability caused by running resource intensive LMM services
on the same nodes as the other control plane services.

Two period jobs are also added, using the new input model generator
scenario, one for cloud8 and another for cloud9. If these jobs prove
to be more stable than the `std-min` jobs used by Gerrit and IBS gating
jobs, the next step will be to switch them to use the `std-lmm` jobs
instead.