Disable publishing for the ci test projects
The ci test projects don't need to be published, all we
care about is the build result. the publishing actually blocks
the project form "completing" which causes timeouts in the ci
run, and we don't need to wait for it.

The script always tried to disable publishing, but the build
service actually ignores that request and enables publishing anyway.
just doing it twice together with a sleep seems to be the magic trick
to let OBS do what we want.