openstack-ardana: include SLES test-updates by default (SCRD-8295)
Change the default behaviour of all Ardana Jenkins jobs
to include SLES test-updates by default.

This is required to align the Ardana and Crowbar CIs
(Crowbar has been doing that for a long time already)
and prevent situations that break the Ardana CI but not
the Crowbar CI, or vice-versa, because of SLES maintenance
updates that are in a `testing` state (i.e. not yet released
but already validated by automated OpenQA tests).

The downside of this change is if there's a SLES package
test-update package that breaks one job, it will break all
jobs - Gerrit CI, IBS gating, periodic regression jobs etc.,
most likely for both Crowbar and Ardana. The upside of this
is that by breaking the CI early, we force everyone to be
proactive instead of re-active and fix potential problems
before candidate SLES package updates are released to customers.

This change also updates the documentation and covers the
use-case of jobs that test the maintenance update scenario:

- jobs that test updating a cloud from the devel cloudsource to
the staging cloudsource (e.g. the cloud8 update gating job)
always install SLES test-updates during the deploy phase
- jobs that test updating a cloud from GM+up to GM+up+test-updates
always install SLES test-updates during the update phase