openstack-crowbar: refactor ses ardana integration
Refactor the SES Ardana integration artifacts (roles,
playbooks, variables) to prepare them to be reused for
Crowbar:
 - rename the `ardana-ses-integration.yml` playbook
 to `cloud-ses-integration.yml`
 - rename the `ses_ardana_integration` role to
 `ses_cloud_integration`
 - rename role variables
 - make the ansible user dependent on cloud type

NOTE: in theory, the refactored playbook might also be
used to provide direct Ceph integration for Crowbar, but
the main purpose of this refactoring is to enable SES
integration (i.e. SES v.2 integration, which uses the Ceph
configuration generated via the SES salt API).