openstack-crowbar: refactor top-level playbooks
Refactor the ansible playbooks and roles such that that every
stage in the Jenkinsfile corresponds to one and only one playbook,
instead of having to call the `deploy-crowbar.yml` playbook multiple
times from the same Jenkinsfile stage.

This change is required to further decouple the back-end pipeline
implementation from the Jenkinsfile definition and also in preparation
of follow-up stages to come (e.g. SES integration), that will require
adding even more back-end logic to these stages.

The following playbooks and roles are added and/or updated by
this refactoring change:
 - `crowbar_setup` role (replaces the previous `crowbar_deploy`
 role) - a simple task wrapping a `qa_crowbarsetup.sh`
 command being executed on the admin node. It is called by the
 following top-level Crowbar playbooks, sometime even multiple times
 from the same playbook, to run relevant `qa_crowbarsetup.sh` commands.
 - `install-crowbar.yml` playbook - installs Crowbar on the admin node.
 - `register-crowbar-nodes.yml` - registers crowbar nodes, same as before
 - `deploy-crowbar.yml` playbook (used to be just a wrapper for the
 `crowbar_deploy` role) - now this is the top-level playbook that deploys
 the Crowbar batch scenario
 - `run-crowbar-tests.yml` - runs tests on the deployed Crowbar cloud.