openstack-crowbar: fix crowbar node order problem
The node ordering that `qa_crowbarsetup.sh` uses to determine
which nodes are controlles, which are computes etc. is not
deterministic, because it's based on sorting the hostnames of the
nodes, which are generated based on the MAC address of their
admin interface.

This change implements a new `custom_crowbar_node_order` environment
variable, which can be used to override the order of nodes that
`qa_crowbarsetup.sh` uses. This variable is then set in the Crowbar
ECP CI so that it corresponds to the node list in the generated heat
stack.

With the correct node order in place, the Crowbar ECP CI no longer
needs to explicitly set up the roles and aliases. It can rely on the
`qa_crowbarsetup.sh` logic, same as mkcloud.