Make sure that after toggling ssl on/off all services are restarted
With the current workflow the periodic chef-client run
might interfere and roll out our keystone toggle changes
partially on the cluster. Make sure everything is in sync
before going with the integration test. We also have to ignore
failures because turning SSL on or off means that the catalog
needs to be updated, OpenStack services needs to switch to
using the new endpoint and be reloaded, Apache needs to be
updated and reloaded, and haproxy needs to be updated and
reloaded. That takes several seconds, but even if it would be
instantaneous, all clients need to switch to a new endpoint.
It would be impossible to do this transparently.

So ensuring that we are not increasing failcounts, we put
the cluster into maintenance mode and then return afterwards
to normal operation.