ipv6: When want_ipv6 is set also set firmware_type to uefi
The want_ipv6 option enables a single stack IPv6 control plane for
crowbar. As such, in order to netboot on vms on an IPv6 admin network
they need to use an efi firmware.

This patch changes the firmware_type to 'uefi' when the want_ipv6
option is enabled.

Also add the want_ipv6 option to the usage/help.