Re-source the crowbar variables after commit
The 'crowbar' command uses the CROWBAR_KEY environment variable to
authenticate. With this upcoming change in crowbar-core[1], the content
of the key is subject to change, so we need to refresh the key between
crowbar commits. This is a simpler path than converting to the
crowbarctl command, which uses the crowbarrc file which does not change.

[1] https://github.com/crowbar/crowbar-core/pull/1802