openstack-crowbar/ardana: exit early in case of failures
Use the `any_errors_fatal` flag when setting up the cloud nodes
to exit as soon as one failure occurs for one of the nodes instead
of continuing to provision the nodes that don't fail up to the end
of the playbook.

This makes it easier to debug one-node ansible failures, because the
last lines in the log will hold the actual failure instead of the
tasks being succesfully executed on the remaining, non-failing nodes.