openstack-crowbar: wait more for nodes to register
The amount of time that it takes for cloud nodes the register
with the admin node in parallel is proportional to the number of
nodes. The number of attempts used for the tasks that check whether
nodes have correctly registered also needs to be proportional to
the number of cloud nodes, to avoid running into timeouts.