openstack-ardana: set number of tempest workers (SCRD-7496)
The current implementation uses the number of CPUs as the number of
tempest workers, this causes some confusion when running tempest tests
on baremetal/virtual deployments or using input models with
standalone/integrated deployers, as some runs faster/slower and
the results differ.

This change sets the number of tempest workers to 4 in all cases, this
value can be changed on demand by overriding the value of
`tempest_run_concurrency`.